Refactor scene_otsu into a modular structure and optimize algorithm to O(T)
