+++ backend/app/scene_otsu/__init__.py
@@ -1,3 +1,19 @@
-from .core import SceneSplitter, SubtitleParser
+from .embedders import (BaseEmbedder, OllamaEmbedder, OpenAIEmbedder,
+                        create_embedder)
+from .parsers import SubtitleParser, scenes_to_srt_string
+from .splitter import SceneSplitter
+from .types import SceneSegment, SubtitleItem
+from .utils import TimestampConverter
 
-__all__ = ["SceneSplitter", "SubtitleParser"]
+__all__ = [
+    "BaseEmbedder",
+    "OpenAIEmbedder",
+    "OllamaEmbedder",
+    "create_embedder",
+    "SubtitleParser",
+    "scenes_to_srt_string",
+    "TimestampConverter",
+    "SceneSplitter",
+    "SubtitleItem",
+    "SceneSegment",
+]
